+++ fj-doc-base/src/test/java/test/org/fugerit/java/doc/base/feature/FeatureConfigTest.java
@@ -1,0 +1,15 @@
+package test.org.fugerit.java.doc.base.feature;
+
+import org.fugerit.java.doc.base.feature.FeatureConfig;
+import org.fugerit.java.doc.base.typehelper.generic.GenericConsts;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+class FeatureConfigTest {
+
+    @Test
+    void isFailWhenElementNotFoundTest() {
+        Assertions.assertEquals(GenericConsts.FAIL_WHEN_ELEMENT_NOT_FOUND_DEFAULT, FeatureConfig.DEFAULT.isFailWhenElementNotFound());
+    }
+
+}